store - change collection update action name, create collection form fields
Feature #14103

Arvados-DCO-1.1-Signed-off-by: Janicki Artur <artur.janicki@contractors.roche.com>
